fix: BUG - Imported folders and documents from G-drive options are limited - EXO-75176
Prior to this fix, cloned Folders and documents from gdrive cannot be managed by users from the Document app, this is because all these items are created using a System session.
This commit change the session to use the current user one so documents wan be managed by all users that have the right permissions, this commit also fix the location where the documents will be cloned.